@@ -22,7 +22,7 @@
     "clsx": "^2.1.0",
     "date-fns": "^3.3.1",
     "embla-carousel-react": "^8.0.0",
-    "next": "14.0.4",
+    "next": "^14.2.3",
     "react": "^18",
     "react-day-picker": "^8.10.0",
     "react-dom": "^18",
